\name{BQJdbcConnectionStringR-package}
\alias{BQJdbcConnectionStringR-package}
\alias{BQJdbcConnectionStringR}
\docType{package}
\title{
\packageTitle{BQJdbcConnectionStringR}
}
\description{
\packageDescription{BQJdbcConnectionStringR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BQJdbcConnectionStringR}
\packageIndices{BQJdbcConnectionStringR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{BQJdbcConnectionStringR}

Maintainer: \packageMaintainer{BQJdbcConnectionStringR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
