% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bqAuthHelper.R
\name{createBQConnectionString}
\alias{createBQConnectionString}
\title{Build a connection string using the JDBC symba driver
https://www.simba.com/products/BigQuery/doc/JDBC_InstallGuide/content/jdbc/config-intro-online.htm}
\usage{
createBQConnectionString(projectId = "", defaultDataset = "",
  authType = 2, jsonCredentialsPath = "", accountEmail = "",
  timeOut = 1000, logLevel = 0, logPath = "", proxyHost = "",
  proxyPort = "", proxyPwd = "", proxyUser = "",
  hostPort = "jdbc:bigquery://https://www.googleapis.com/bigquery/v2:443")
}
\arguments{
\item{projectId}{This is the project where you will connect and the submitted SQL queries jobs will be executed. This is not neccesarilly the project where the datasets you are interested lives, but rather the project you have permissions to execute jobs.}

\item{defaultDataset}{This is the dataset where tables referenced on SQL code without a project or dataset id in the name will be created.}

\item{authType}{0 to use a service account and 2 (default) to use a pregenerated refresh token.}

\item{jsonCredentialsPath}{This is the path with the json credentials. It can be from the pregenerated refresh token
or the service account}

\item{accountEmail}{This is the account email for the service account if authType = 0}

\item{timeOut}{The length of time, in seconds, that the driver waits for a query to retrieve the results of an executed job}

\item{logLevel}{0 for desabling 6 to log everything. Details here https://www.simba
.com/products/BigQuery/doc/JDBC_InstallGuide/content/jdbc/options/loglevel.htm}

\item{logPath}{The full path to the folder where the driver saves log files when logging is enabled.}

\item{proxyHost}{The IP address or host name of your proxy server.}

\item{proxyPort}{The listening port of your proxy server.}

\item{proxyPwd}{The password, if needed, for proxy server settings.}

\item{proxyUser}{The user name, if needed, for proxy server settings.}

\item{hostPort}{Host and port the driver is connecting. This parameter usually does not change and its default value work}
}
\description{
Build a connection string using the JDBC symba driver
https://www.simba.com/products/BigQuery/doc/JDBC_InstallGuide/content/jdbc/config-intro-online.htm
}
\details{
This is a helper function to ease the process of creating a connection string for the JDBC BigQuery Driver.
}
