\name{BQJdbcConnectionStringR-package}
\alias{BQJdbcConnectionStringR-package}
\alias{BQJdbcConnectionStringR}
\docType{package}
\title{
\packageTitle{BQJdbcConnectionStringR}
}
\description{
\packageDescription{BQJdbcConnectionStringR}
}
\details{


The DESCRIPTION file:
\packageDESCRIPTION{BQJdbcConnectionStringR}
\packageIndices{BQJdbcConnectionStringR}

}
\author{
\packageAuthor{BQJdbcConnectionStringR}


